+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * added microseconds in Timestamp type
 
 ## 1.8.2
+
 * fixed discovery on exception
 * fixed logger in EnvironCredentials