+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## 1.10.0
 * changed level of update token log record from info to debug
 * created refresh token ratio parameter